add test plans api
